Card selections display on all users' screens
To initiate interactivity, navigate to /server in a bash terminal and run npm start. After this, when a card is selected in one window the username and vote will appear for everyone in the results section. Still need to connect this to the vote average function and the submit button.